Remove the NumberParser and place new one in. Add support for disabling Symbols.
